Huge revamping: introduced plugin-based definitions of challenges
Now each challenge is self-contained in a .py script within the
"challenges" subdir.
Loading of challenges happens based on a TOML configuration file.

Moreover, a "put-the-flag" like challenge has been introduced for the
echo one word challenge. This requires to store in the state the plain
text word so that the "current" command remains consistent.